Define the signup route to add a new user
Create a basic user model with a Mongoose schema. Check if request data
is already present in the database. Generate a hashed password & create
a new user along with a JWT token and cookie. Cleanup eslint and ts
config files. Add comments for third party middleware being used with the
main app.