Set up Zod for validation & Pino for logging
Create User resource and associated model, service, and controller to
create a new user. The controller communicates with the user service,
which then communicates with the Mongo database. Also, set up pino
logger which outputs to both the console and a server log file. Add an
npm script to pull latest frontend code in the subtree. Use Zod to
validate HTTP request data.