Add session model, service, and controller
Create a session for the user by authenticating and generating access
and refresh tokens using JWT. Deserialize the user for all routes.
Create a middleware to check if the user is required to be logged in
to access a particular route. Update the error handler to handle Mongo
errors properly. Ensure the testcase for user creation failure passes.
Use RSA public/private keys for JWT as its more secure. Add a file that
documents HTTP requests for the auth routes.